Сделал небольшую структуру проекта согласно пункту 3 Перенёс константы и генераторы в директорию mock-data
