--- js/data/constants.ts
@@ -1,11 +1,0 @@
-export const enum Constants {
-	MIN_ID = 1,
-	MIN_PHOTO_ADDRESS = 1,
-	MAX_PHOTO_ADDRESS = 25,
-	MIN_LIKES = 15,
-	MAX_LIKES = 200,
-	MIN_COMMENTS = 0,
-	MAX_COMMENTS = 30,
-	MIN_AVATAR_ID = 1,
-	MAX_AVATAR_ID = 6,
-}
